chore: fixing tenant id to int and squash migrations

+++ srcs/infra/Migrator/Migrations/Srv/Identity/MicroserviceSrvIdentityDbContextModelSnapshot.cs
@@ -184,8 +184,8 @@ namespace datntdev.Microservice.Infra.Migrator.Migrations.Srv.Identity
                     b.Property<long>("UserId")
                         .HasColumnType("bigint");
 
-                    b.Property<long>("TenantId")
-                        .HasColumnType("bigint");
+                    b.Property<int>("TenantId")
+                        .HasColumnType("int");
 
                     b.Property<DateTime?>("CreatedAt")
                         .HasColumnType("datetime2");
